+++ hercules-ci-cnix-expr/test/data/simple-flake/flake.nix
@@ -1,0 +1,7 @@
+{
+  description = "A very simple flake";
+  inputs = { };
+  outputs = { ... }: {
+    lib.greeting = "hello flake";
+  };
+}