+++ hercules-ci-cnix-expr/CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+ - `getLocalFlake` using either flake-compat or Nix 2.4pre
+
 ## 0.2.0.0 - 2021-06-22
 
 ### Added
 
  - nixUnstable compatibility with flag `nix-2_4`
+
+### Changed
+
+ - Flakes are enabled during `init` on Nix 2.4pre
 
 ## 0.1.0.0 - 2021-03-07